create Meals.js file and separated the  Home.js file from the App.js file
