+++ frontend/kitchen-served-fe-app/package.json
@@ -10,6 +10,7 @@
     "lucide-react": "^0.453.0",
     "react": "^18.3.1",
     "react-dom": "^18.3.1",
+    "react-router-dom": "^6.27.0",
     "react-scripts": "5.0.1",
     "tailwindcss": "^3.4.14",
     "web-vitals": "^2.1.4"
